+++ src/Repository/TaskRepositoryInterface.php
@@ -5,8 +5,9 @@ namespace Bupy7\Queue\Repository;
 interface TaskRepositoryInterface
 {
     /**
-     * Returns entities has status Task::STATUS_RUN or Task::STATUS_ERROR for run.
-     * @return \Bupy7\Queue\Entity\Task[]
+     * Returns entities has status \Bupy7\Queue\Entity\TaskInterface::STATUS_RUN or
+     * \Bupy7\Queue\Entity\TaskInterface::STATUS_ERROR for run.
+     * @return \Bupy7\Queue\Entity\TaskInterface[]
      */
     public function findForRun(int $limit = null): array;
 }