+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 zf-queue
 ========
 
+1.0.1 [2017-11-22]
+------------------
+
+- Fix: Skiped the param in the `Bupy7\Queue\Manager\EntityManagerInterface::flush()` method.
+- Enh: Small enhancements.
+- Add: `executed` event in the `Bupy7\Queue\Service\QueueService`. 
+
 1.0.0 [2017-09-13]
 ------------------